Changed miles convert function to allow for float values

+++ main.py
@@ -15,7 +15,7 @@ km_label = Label(text="Km")
 km_label.grid(column=2, row=1)
 
 def convert_miles_to_km():
-  km = int(miles_input.get())*1.60934
+  km = float(miles_input.get())*1.60934
   km_value.config(text=km)
 
 calculate_button = Button(text="Calculate", command=convert_miles_to_km)
